+++ src/context/kemaisession.h
@@ -7,6 +7,7 @@
 #include "client/kimaiCache.h"
 #include "client/kimaiapi.h"
 #include "client/kimaiclient.h"
+#include "monitor/kimaiEventsMonitor.h"
 
 namespace kemai {
 
@@ -15,7 +16,7 @@ class KemaiSession : public QObject
     Q_OBJECT
 
 public:
-    KemaiSession(std::shared_ptr<KimaiClient> kimaiClient);
+    KemaiSession(const std::shared_ptr<KimaiClient>& kimaiClient);
     ~KemaiSession() override;
 
     std::shared_ptr<KimaiClient> client() const;
@@ -48,16 +49,13 @@ private:
     void requestTimeSheetConfig();
     void requestPlugins();
 
-    void setCurrentTimeSheet(const TimeSheet& timeSheet);
-    void clearCurrentTimeSheet();
-
     std::shared_ptr<KimaiClient> mKimaiClient;
     KimaiCache mKimaiCache;
+    KimaiEventsMonitor mKimaiMonitor;
     QVersionNumber mKimaiVersion;
     Plugins mPlugins;
     User mMe;
     TimeSheetConfig mTimeSheetConfig;
-    std::optional<TimeSheet> mCurrentTimeSheet;
 };
 
 } // namespace kemai